Test: 平滑曲线重绘

+++ UI/EnterAffectiveCloudUI/View/Report3/AffectiveCharts3Pressure.swift
@@ -64,7 +64,7 @@ public class AffectiveCharts3Pressure: AffectiveCharts3LineCommonView {
         return self
     }
     
-    public override func setData(_ array: [Double]) -> Self {
+    public override func setData(_ array: [Int]) -> Self {
         guard array.count > 0 else {return self}
         dataSorce.removeAll()
         separateY.removeAll()
@@ -159,11 +159,11 @@ public class AffectiveCharts3Pressure: AffectiveCharts3LineCommonView {
     }
     
     public override func build() {
-        let invalidData = 5.0
+        let invalidData = 5
         var yVals: [ChartDataEntry] = []
         var data:LineChartData!
         if theme.style == .session {
-            var initValue = 0.0 //初始数据
+            var initValue = 0 //初始数据
             var initIndex = 0 //开始有值索引位置
             for i in stride(from: 0, to: dataSorce.count, by: sample) {
                 let value = dataSorce[i]
